sites: developed /en/ landing page
Thanks to Cory's written script, we can now develop the landing page
marked as the first prototype for now. Some non-critical components
like "Teams" are intentionally left out due to a deadline and shall
be implemented in a separate commit. Hence, let's do this.

This patch develops /en/ landing page in sites/ directory.

Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>