sites: un-hide the "why us" section header
Sibert provided his feedback that the page is missing "why us"
section where we actually have it. This means the current design
failed so badly that people are not intuitively getting it. Hence,
let's un-hide it and restore to what we know it works.

This patch un-hides the "why us" section header in sites/
directory.

Co-authored-by: Sibertius <sibertius@gmail.com>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>