sites: fixed corygalynaCARD's non-centered thumbnails on small screen
Appearently, the images presented by corygalynaCARD's .thumbnail
class is not displaying properly (non-centered and always on the left).
This is caused by browser treating 'picture > img' not the same as
'picture'. Hence, we need to fix/workaround it.

This patch fixes corygalynaCARD's non-centered thumbnails on small
screen in sites/ directory.

Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>