sites: first refactored for non-technical content
Sibert and Cory both agreed independently that the current contents
are too technical and not friendly to non-tech individuals. Hence,
we will have to refactor it with simple sentences. Let's do the
first refactoring.

This patch refactors the entire content in /en page to be more
non-tech people friendly in sites/ directory.

Co-authored-by: Sibertius <sibertius@gmail.com>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>