sites: 2nd content simplification re-factoring efforts
After getting a pass from a granny test and feedbacks from Sibert,
we have to remove the remaining acronyms and use simple English.
The goal is not to make visitor to Google search terms when
experiencing the entire journey. Hence, let's refactor it again.

This patch refactors the /en content with simple english in sites/
directory for the 2nd time.

Co-authored-by: Sibertius <sibertius@gmail.com>
Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>