sites: temporarily configures the 404 redirection to a single page
Since we don't have time to mess around with 404 pages, we should
configure it to head back towards the available landing page which
is /en as of this commit. Hence, let's reconfigure it.

This patch temporarily configures the 404 redirection to a single
page (/en) and making sure the sites/ are working fine.

Co-authored-by: Galyna, Cory <cory.galyna@zoralab.com>
Co-authored-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>
Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>